Pushed v1.58 code (mega-commit again, sorry)
- Added support for loading FLAC samples
- The IFF sample saver was writing somewhat broken files for non-looped samples
- The "UPSAMP" button in Edit Op. was showing the wrong status text
- Small code refactor for how the song is timed/ticked. Longer song-to-WAV renders may have a filesize change of a few bytes, but this is actually more correct.
- Increased number of precision bits for song BPM, playback time counter and audio->video syncing. This is a minor change and does very little in practice as the previous precision was already quite good.
- Audio/video sync timestamps are reset every half an hour to prevent possible sync drifting after several hours of playing a song without a single song stop (resets timestamp) in-between
- The HPC timers are reset every half an hour instead of every hour. The video (if no vsync) and the scopes use these timers. This hopefully fixes potential timing drifting issues that were still lingering.
- Code refactoring for module/sample loaders
- Windows/macOS: Updated SDL to v2.26.5

+++ src/pt2_paula.h
@@ -22,7 +22,7 @@ void paulaDisableFilters(void); // disables low-pass & high-pass filters ("LED" 
 
 int8_t *paulaGetNullSamplePtr(void);
 
-void paulaWriteByte(uint32_t address, uint16_t data8);
+void paulaWriteByte(uint32_t address, uint8_t data8);
 void paulaWriteWord(uint32_t address, uint16_t data16);
 void paulaWritePtr(uint32_t address, const int8_t *ptr);
